requirements.txt
absl-py==2.1.0
accelerate==0.20.3
adal==1.2.7
adjustText==1.0.4
aiohttp==3.8.6
aiosignal==1.3.1
alabaster==0.7.13
altair==4.2.2
anyio==3.7.1
argcomplete==2.1.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
asynctest==0.13.0
atomicwrites==1.4.1
attrs==23.2.0
Automat==22.10.0
azure-common==1.1.28
azure-core==1.29.4
azure-graphrbac==0.61.1
azure-identity==1.7.0
azure-mgmt-authorization==2.0.0
azure-mgmt-containerregistry==10.0.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==10.0.0
azure-mgmt-resource==21.1.0
azure-mgmt-storage==20.0.0
azure-storage-blob==12.10.0
azureml-core==1.43.0
azureml-dataprep==4.0.4
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.6.3
azureml-dataset-runtime==1.43.0
Babel==2.14.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.8.0
benepar==0.2.0
black==22.1.0
bleach==3.2.1
blinker==1.6.3
blis==0.4.1
cachetools==5.3.3
catalogue==1.0.0
category-encoders==2.3.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
constantly==15.1.0
contextlib2==21.6.0
coverage==7.2.7
cryptography==37.0.4
cssselect==1.2.0
cycler==0.10.0
cymem==2.0.4
Cython==0.29.28
dataclasses==0.6
decorator==5.1.1
defusedxml==0.6.0
distro==1.8.0
docker==5.0.3
docopt==0.6.2
docutils==0.19
dotnetcore2==3.1.23
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.3.1/en_core_web_lg-2.3.1.tar.gz
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastjsonschema==2.18.1
filelock==3.0.12
flake8==3.7.9
Flask==2.0.1
flatbuffers==24.3.25
fpdf==1.7.2
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-image-source-search==0.0.2
google-pasta==0.2.0
great-expectations==0.16.13
grpcio==1.62.1
h5py==3.8.0
huggingface-hub==0.16.4
humanfriendly==10.0
hyperlink==21.0.0
idna==2.10
imagesize==1.4.1
importlib-metadata==5.2.0
incremental==22.10.0
inference-schema==1.1.0
iniconfig==2.0.0
ipykernel==5.3.4
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.1
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.17.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.0
joblib==0.17.0
jsonnet-binary==0.17.0
jsonpatch==1.33
jsonpickle==2.2.0
jsonpointer==2.0
jsonschema==4.0.0
jupyter==1.0.0
jupyter-client==7.4.9
jupyter-console==6.2.0
jupyter-core==4.12.0
jupyter-server==1.24.0
jupyterlab-pygments==0.1.2
keras==2.11.0
kiwisolver==1.3.1
knack==0.9.0
libclang==18.1.1
lightgbm==3.1.1
lightning-utilities==0.9.0
llvmlite==0.39.1
lxml==4.5.0
makefun==1.15.1
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.6.1
mdurl==0.1.2
MechanicalSoup==1.3.0
mergedeep==1.3.0
missingno==0.4.2
mistune==3.0.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
mlxtend==0.19.0
msal==1.24.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
multidict==6.0.4
murmurhash==1.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.5.1
nbconvert==7.6.0
nbformat==5.8.0
ndg-httpsclient==0.5.1
nervaluate==0.1.7
nest-asyncio==1.5.8
networkx @ file:///tmp/build/80754af9/networkx_1633639043937/work
nltk==3.4.5
notebook==6.5.6
notebook-shim==0.2.3
num2words==0.5.13
numba @ file:///C:/b/abs_e53pp2e4k7/croot/numba_1670258349527/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1653574840943/work
oauthlib==3.2.2
open3d==0.10.0.0
opencv-python==4.3.0.36
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.4.3
paramiko==2.12.0
parsel==1.8.1
parso==0.7.1
pathspec==0.11.2
patsy==0.5.3
pdflatex==0.1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
pkginfo==1.9.6
plac==1.1.3
platformdirs==3.10.0
plotly==5.0.0
pluggy==0.13.1
portalocker==2.7.0
preshed==3.0.4
prometheus-client==0.8.0
prompt-toolkit==3.0.8
Protego==0.3.0
protobuf==3.9.2
psutil==5.9.6
psynlig==0.2.2.dev0
ptyprocess==0.6.0
py==1.11.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.5.0
pycparser==2.20
pydantic==1.10.13
pydeck==0.8.1b0
pydevd-pycharm==212.4746.96
PyDispatcher==2.0.7
pyflakes==2.1.1
Pygments==2.16.1
PyJWT==2.8.0
PyLaTeX==1.4.1
Pympler==1.0.1
PyNaCl==1.5.0
pyngrok==5.1.0
pynndescent==0.5.11
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==6.1.2
pytest-check==1.0.1
pytest-cov==2.11.1
pytest-mock==3.6.1
python-dateutil==2.8.2
python-dev-settings==1.0.12
pytorch-lightning==1.9.5
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pywin32==227
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==4.7.7
QtPy==1.9.0
queuelib==1.6.2
regex==2019.6.8
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
rfc3987==1.3.8
rich==13.6.0
rsa==4.9
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.8
sacremoses==0.0.43
safetensors==0.4.0
scikit-learn==0.22.2.post1
scipy==1.7.3
Scrapy==2.9.0
seaborn==0.11.0
SecretStorage==3.3.3
Send2Trash==1.8.2
sentencepiece==0.1.99
service-identity==21.1.0
simplejson==3.17.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
spacy==2.3.2
sphinx==5.3.0
sphinx-gallery==0.14.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==1.0.4
statsmodels==0.12.2
streamlit==1.23.1
strict-rfc3339==0.7
svgling==0.4.0
svgwrite==1.4.3
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.9.1
testpath==0.4.4
textblob==0.17.1
thinc==7.4.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tldextract==4.0.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torch-struct==0.5
torchmetrics==0.11.4
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.30.2
twisted==23.8.0
twisted-iocpsupport==1.0.4
typed-ast==1.5.5
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
umap-learn==0.5.5
urllib3==1.26.1
validators==0.20.0
w3lib==2.1.2
wasabi==0.8.0
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.1
xgboost==1.3.3
xlrd==1.1.0
yarl==1.9.2
zipp==3.15.0
zope.interface==6.2
